+++ pages/api/routes/lookbook/[id].js
@@ -1,9 +1,16 @@
 import { deleteImage } from '../../controllers/lookbook';
+import { checkAuth } from '../../middlewares/checkAuth';
 
-export default function handler(req, res) {
+export default async function handler(req, res) {
   switch (req.method) {
     case 'DELETE':
-        deleteImage(req, res)
+        const user = await checkAuth(req, res);
+        if (!user || user.role !== 'admin') {
+          return res.status(403).json({ error: `Нет прав` });
+        } else {
+          deleteImage(req, res)
+        }
+        // deleteImage(req, res)
         break;
     default:
         res.status(405).json({ error: `Недопустимый метод` });
